+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Only create media directory in export when needed
 - Export view model exposes more info in error message
 - Export uses better timestamp formats
+- Upgrade Trixnity to 4.6.1
 
 ### Deprecated